Remove duplicate types
