+++ front/app/containers/Admin/projects/project/nativeSurvey/FormResults/FormResultsQuestion/index.tsx
@@ -4,7 +4,7 @@ import { Box, Title } from '@citizenlab/cl2-component-library';
 import { snakeCase } from 'lodash-es';
 import { Locale } from 'typings';
 
-import { Result } from 'api/survey_results/types';
+import { ResultUngrouped } from 'api/survey_results/types';
 
 import T from 'components/T';
 
@@ -15,7 +15,7 @@ import MultipleChoice from './MultipleChoice';
 import PointLocationQuestion from './PointLocationQuestion';
 import TextQuestion from './TextQuestion';
 
-type FormResultsQuestionProps = Result & {
+type FormResultsQuestionProps = ResultUngrouped & {
   questionNumber: number;
   locale: Locale;
   totalSubmissions: number;
